requirements.txt
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
fonttools==4.61.1
importlib_resources==6.5.2
Jinja2==3.1.6
kiwisolver==1.4.9
loguru==0.7.3
MarkupSafe==3.0.3
matplotlib==3.10.8
numpy==2.2.6
packaging==26.0
pandas==2.3.3
pillow==12.1.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
sweetviz==2.3.1
tqdm==4.67.3
tzdata==2025.3
win32_setctime==1.2.0
